+++ src/main/java/com/cronolytics/api/repository/IEmpresaRepository.java
@@ -4,7 +4,7 @@ import com.cronolytics.api.entity.Empresa;
 import com.cronolytics.api.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IEmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface IEmpresaRepository extends JpaRepository<Empresa, Integer> {
     Empresa findByEmail(String email);
     Empresa findByEmailAndSenha(String email, String senha);
